==> 2025_12_manuscript_analysis/src/04_plot_draft_pangenome_comparison.Rmd <==
---
title: "pure_assembly_data"
author: "Julian A Paganini"
date: "2025-12-12"
output: html_document
---

---
title: "Pangenome plot comparison"
author: "Julian A Paganini"
date: "2025-12-03"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(ggplot2)
library(tidyr)
library(broom)
library(dplyr)
library(readxl)
library(reshape2)
library(tidyverse)
library(viridis)
library(stringr)
library(patchwork)
```

##1. Import data and Wrangle
###1. Import degrees, lengths and colours

```{r}
#1. Import all assembly genome data
complete_lengths_subgraphs_id <- read.csv("../results/pangenome_parsed/02_all_assembly/length_results.csv")
complete_degree_id <- read.csv("../results/pangenome_parsed/02_all_assembly/degree_results.csv")
#combine information
all_complete_data<-left_join(complete_lengths_subgraphs_id,complete_degree_id,by=c("Tool","node_id"="Node"))
all_complete_data<-all_complete_data%>% filter(Tool!="pangraph")
#change names
all_complete_data <- all_complete_data %>%
  mutate(
    Tool = recode(
      Tool,
      # "pangraph" = "Pangraph",
      "panaroo" = "Panaroo",
      "ppanggolin" = "PPanGGOLiN",
      "ggcaller" = "ggCaller",
      "bifrost" = "Bifrost",
      "cuttlefish" = "Cuttlefish"
      # add more as needed
    )
  )

#Re-order tools
all_complete_data$Tool <- factor(
  all_complete_data$Tool,
  levels = c("Panaroo",
             "PPanGGOLiN",
             "ggCaller",
             "Bifrost",
             "Cuttlefish")
             # "Pangraph")
)

all_complete_data$node_id <- gsub("'", "", all_complete_data$node_id)


#Import node abundances
bifrost_abundances_filt<- read.csv("../../2024_01_pangenome_constructions/results/pangenome_graph/pure_assembly/abundances/bifrost_unitig_id_abundance.tsv", sep="\t")
bifrost_abundances_filt<-bifrost_abundances_filt[,c(177:179)]
#complete for the rest
names(bifrost_abundances_filt)<-c("node_id","n_isolates","fraction_isolates")
bifrost_abundances_filt$Tool<-"Bifrost"
#
cuttlefish_abundances_filt<- read.csv("../../2024_01_pangenome_constructions/results/pangenome_graph/pure_assembly/abundances/cuttlefish_unitig_id_abundance.csv")
cuttlefish_abundances_filt<-cuttlefish_abundances_filt[,c(1:3)]
names(cuttlefish_abundances_filt) <- c("node_id","n_isolates","fraction_isolates")
cuttlefish_abundances_filt$Tool <- "Cuttlefish"

ggcaller_abundances_filt<- read.csv("../../2024_01_pangenome_constructions/results/pangenome_graph/pure_assembly/abundances/ggcaller_gene_presence_absence_abundance.tsv", sep="\t")
ggcaller_abundances_filt<-ggcaller_abundances_filt[,c(1,177,178)]
names(ggcaller_abundances_filt) <- c("node_id","n_isolates","fraction_isolates")
ggcaller_abundances_filt$Tool <- "ggCaller"

panaroo_abundances_filt<- read.csv("../../2024_01_pangenome_constructions/results/pangenome_graph/pure_assembly/abundances/panaroo_gene_presence_absence_abundance.tsv", sep="\t")
panaroo_abundances_filt<-panaroo_abundances_filt[,c(1,177,178)]
names(panaroo_abundances_filt) <- c("node_id","n_isolates","fraction_isolates")
panaroo_abundances_filt$Tool <- "Panaroo"

ppanggolin_abundances_filt<- read.csv("../../2024_01_pangenome_constructions/results/pangenome_graph/pure_assembly/abundances/ppanggolin_gene_id_abundance.csv")
#add number
ppanggolin_abundances_filt$Gene <- 1:nrow(ppanggolin_abundances_filt)
ppanggolin_abundances_filt<-ppanggolin_abundances_filt[,c(1,3)]
names(ppanggolin_abundances_filt) <- c("node_id","n_isolates")
ppanggolin_abundances_filt$fraction_isolates <- round(ppanggolin_abundances_filt$n_isolates / 175, 9)
ppanggolin_abundances_filt$Tool <- "PPanGGOLiN"


all_abundances <- rbind(
  bifrost_abundances_filt[, c("node_id","n_isolates","fraction_isolates","Tool")],
  cuttlefish_abundances_filt[, c("node_id","n_isolates","fraction_isolates","Tool")],
  ggcaller_abundances_filt[, c("node_id","n_isolates","fraction_isolates","Tool")],
  panaroo_abundances_filt[, c("node_id","n_isolates","fraction_isolates","Tool")],
  ppanggolin_abundances_filt[, c("node_id","n_isolates","fraction_isolates","Tool")])



all_complete_data<-left_join(all_complete_data,all_abundances,by=c("Tool","node_id"))


```

#2. Figure 3 - Differences in basic graph structure

```{r fig.width=7,fig.height=4, dpi=300}
agd<-read.csv("../results/pangenome_parsed/06_aggregated_data/basic_information.csv", check.names = FALSE, stringsAsFactors = FALSE)

names(agd) <- c(
  "Tool",
  "node_count",
  "edge_count",
  "subgraph_count",
  "length_median",
  "length_q1",
  "length_q3",
  "pct_complete",
  "method"
)

agd <- agd %>%
  mutate(
    Tool = recode(
      Tool,
      # "pangraph" = "Pangraph",
      "panaroo" = "Panaroo",
      "ppanggolin" = "PPanGGOLiN",
      "ggcaller" = "ggCaller",
      "bifrost" = "Bifrost",
      "cuttlefish" = "Cuttlefish"
      # add more as needed
    )
  )


agd <- agd %>%
  mutate(
    pct_complete = as.numeric(gsub("%", "", pct_complete)),
    pct_complete = recode(
      pct_complete,
      `10` = 11,
      `90` = 89,
      .default = pct_complete
    )
  ) %>%
  filter(Tool != "pangraph")

# 2) Keep only 11, 50, 89 for plotting, but keep 100 for baseline computation
agd_keep <- agd %>%
  filter(pct_complete %in% c(0, 11, 50, 89, 100))

# 3) Long format for the three metrics
long <- agd_keep %>%
  pivot_longer(
    cols = c(node_count, edge_count, subgraph_count),
    names_to = "metric",
    values_to = "value"
  )

# 4) Compute baseline (100%) per Tool + metric, then diff vs baseline
diff_df <- long %>%
  group_by(Tool, metric) %>%
  mutate(baseline_100 = value[pct_complete == 100][1]) %>%
  ungroup() %>%
  filter(pct_complete != 100) %>%
  mutate(
    abs_change = value - baseline_100,
    pct_change = 100 * (value - baseline_100) / baseline_100,
    direction = ifelse(pct_change >= 0, "Up", "Down"),
    pct_complete = factor(pct_complete, levels = c(0, 11, 50, 89))
  )

diff_df<-diff_df[,c(1,5,7:12)]

diff_df <- diff_df %>%
  mutate(
    metric = recode(
      metric,
      node_count = "Node count",
      edge_count = "Edge count",
      subgraph_count = "Subgraph count"
    )
  )

#write.csv(diff_df, '../results/Supplementary_table_S6.csv',  row.names = FALSE)

# 5) Plot
ggplot(diff_df, aes(x = pct_complete, y = pct_change, fill = direction)) +
  geom_hline(yintercept = 0, linetype = "dotted", linewidth = 0.6) +
  geom_col(width = 0.75) +
  scale_fill_manual(values = c("Up" = "red3", "Down" = "steelblue3")) +
  facet_grid(metric ~ Tool, scales = "free_y") +
  labs(
    x = "Percentage of complete genomes",
    y = "Percent change vs 100% complete graph",
    fill = NULL
  ) +
  theme_bw() +
  theme(
    strip.text = element_text(face = "bold", size=14),
    axis.text.x = element_text(size = 12),
    axis.text.y = element_text(size = 12),
    axis.title.x = element_text(size = 14,face = "bold"),
    axis.title.y = element_text(size = 14,face = "bold")
  )
```
#3. Node-degree distribution per tool
```{r fig.width=7, fig.height=4, dpi=300}
bins <- seq(0, 1, by = 0.05)

df <- all_complete_data %>%
  filter(
    Degree >= 0, Degree <= 10,
    fraction_isolates >= 0, fraction_isolates <= 1
  ) %>%
  mutate(
    degree_bin = Degree,
    frac_bin_raw = cut(fraction_isolates, breaks = bins, include.lowest = TRUE),
    frac_bin = factor(
      frac_bin_raw,
      labels = paste0(seq(0, 95, 5), "%–", seq(5, 100, 5), "%")
    )
  )

# Count per bin per Tool
df_bins <- df %>%
  group_by(Tool, degree_bin, frac_bin) %>%
  summarise(count = n(), .groups = "drop")

# TOTAL COUNTS per tool (normalization)
totals <- df %>%
  group_by(Tool) %>%
  summarise(total_nodes = n())

# Normalize by total # nodes of that Tool
df_bins <- df_bins %>%
  left_join(totals, by = "Tool") %>%
  mutate(count_norm = count / total_nodes)


df_bins$Tool <- factor(
  df_bins$Tool,
  levels = c("Panaroo",
             "PPanGGOLiN",
             "ggCaller",
             "Bifrost",
             "Cuttlefish")
)

#write.csv(df_bins, '../results/TMP_node_abundance.csv',  row.names = FALSE)

# Plot
ggplot(df_bins, aes(x = degree_bin, y = frac_bin, fill = count_norm)) +
  geom_tile() +
  scale_fill_viridis_c(
    name = "Normalized\nnode count",
    limits = c(0, NA)
  ) +
  scale_x_continuous(breaks = 0:10, name = "Node degree") +
  ylab("Fraction of isolates") +
  facet_wrap(~Tool) +
  theme_bw() +
  theme(
    axis.title.x = element_text(size = 14, face = "bold"),
    axis.text.x = element_text(size = 10),
    axis.title.y = element_text(size = 14, face = "bold"),
    axis.text.y = element_text(size = 10),
    panel.grid = element_blank(),
    strip.text = element_text(size=12,face = "bold")
  )
```

##3.1 Supplementary Figure S2 - Differences in node-degree distribution 

```{r}
complete_genome_node_abundance<- read.csv("../results/IMPORTANT_complete_node_degree_normalized_tool.csv")
complete_genome_node_abundance<-complete_genome_node_abundance%>% filter(Tool!="Pangraph")
colnames(complete_genome_node_abundance)[6] <- "count_norm_baseline"
colnames(complete_genome_node_abundance)[5] <- "total_nodes_baseline"
colnames(complete_genome_node_abundance)[4] <- "count_baseline"

difference_abundance_data<-full_join(complete_genome_node_abundance,df_bins,by=c("Tool","degree_bin","frac_bin"))
difference_abundance_data[is.na(difference_abundance_data)] <- 0

difference_abundance_data$delta_pp_bin<-100 * (difference_abundance_data$count_norm - difference_abundance_data$count_norm_baseline)

#write.csv(difference_abundance_data, '../results/TMP_DIFF_abundance.csv',  row.names = FALSE)

deg_diff <- difference_abundance_data %>%
  group_by(Tool, degree_bin) %>%
  summarise(
    frac_baseline = sum(count_norm_baseline, na.rm = TRUE),
    frac_draft    = sum(count_norm,           na.rm = TRUE),
    .groups = "drop"
  ) %>%
  mutate(
    delta_pp = 100 * (frac_draft - frac_baseline)
  )


deg_diff$Tool <- factor(
  deg_diff$Tool,
  levels = c("Panaroo",
             "PPanGGOLiN",
             "ggCaller",
             "Bifrost",
             "Cuttlefish")
)


#write.csv(deg_diff, '../results/TMP_all_degree_diffrence.csv',  row.names = FALSE)

ggplot(deg_diff, aes(x = degree_bin, y = delta_pp)) +
  geom_hline(yintercept = 0, linetype = "dashed") +
  geom_line() +
  geom_point() +
  facet_wrap(~ Tool, ncol = 3) +
  scale_x_continuous(breaks = 0:10) +
  labs(
    x = "Node degree",
    y = "Change in node fraction (draft – complete, percentage points)"
  ) +
  theme_bw()+
  theme(
    panel.grid.minor = element_blank()
  )


```



##3.2 node abundance an isolate-frequency (Degree-normalized)

```{r fig.width=9, fig.height=5, dpi=300}

# Normalized - total nodes per Tool & degree
totals_deg <- df %>%
  group_by(Tool, degree_bin) %>%
  summarise(total_deg = n(), .groups = "drop")

df_bins_node <- df_bins %>%
  left_join(totals_deg, by = c("Tool", "degree_bin")) %>%
  mutate(count_norm = count / total_deg)


df_bins_node$Tool <- factor(
  df_bins_node$Tool,
  levels = c("Panaroo",
             "PPanGGOLiN",
             "ggCaller",
             "Bifrost",
             "Cuttlefish")
)


# Plot
ggplot(df_bins_node, aes(x = degree_bin, y = frac_bin, fill = count_norm)) +
  geom_tile() +
  scale_fill_viridis_c(
    name = "Node fraction\n(Degree normalized)",
    limits = c(0, NA)
  ) +
  scale_x_continuous(breaks = 0:10, name = "Node degree") +
  ylab("Fraction of isolates") +
  facet_wrap(~Tool) +
  theme_bw() +
  theme(
    axis.title.x = element_text(size = 16, face = "bold"),
    axis.text.x = element_text(size = 12),
    axis.title.y = element_text(size = 16, face = "bold"),
    axis.text.y = element_text(size = 10),
    panel.grid = element_blank(),
    strip.text = element_text(size=16,face = "bold")
  )

```

##3.4 Difference in node abudannce (draft vs complete) / normalize by node-degree 

```{r fig.width=9, fig.height=6, dpi=300}

complete_genome_degree_abundance<- read.csv("../results/Supplementary_table_S5.csv")
complete_genome_degree_abundance<-complete_genome_degree_abundance%>% filter(Tool!="Pangraph")
colnames(complete_genome_degree_abundance)[8] <- "total_count_norm_baseline"
colnames(complete_genome_degree_abundance)[7] <- "total_nodes_baseline"
colnames(complete_genome_degree_abundance)[6] <- "count_norm_baseline"
colnames(complete_genome_degree_abundance)[5] <- "total_deg_baseline"
colnames(complete_genome_degree_abundance)[4] <- "count_baseline"

node_degree_abundance_difference<-full_join(complete_genome_degree_abundance,df_bins_node,by=c("Tool","degree_bin","frac_bin"))
node_degree_abundance_difference<- node_degree_abundance_difference %>% mutate(total_count_norm = count / total_nodes)

node_degree_abundance_difference[is.na(node_degree_abundance_difference)] <- 0

#calculate the difference in node-normalized fashiones
node_degree_abundance_difference$delta_pp_node = round(100 * (node_degree_abundance_difference$count_norm - node_degree_abundance_difference$count_norm_baseline),2)

#calculate the difference in whole-graph normalized fasion
node_degree_abundance_difference$delta_pp_total= round(100 * (node_degree_abundance_difference$total_count_norm - node_degree_abundance_difference$total_count_norm_baseline),2)





frac_levels <- paste0(
  seq(0, 95, by = 5), "%–",
  seq(5, 100, by = 5), "%"
)

node_degree_abundance_difference <- node_degree_abundance_difference %>%
  mutate(
    frac_bin = factor(frac_bin, levels = frac_levels, ordered = TRUE)
  )


#write.csv(node_degree_abundance_difference, '../results/TMP_bin_degree_diffrence.csv',  row.names = FALSE)

node_degree_abundance_difference$Tool <- factor(
  node_degree_abundance_difference$Tool,
  levels = c("Panaroo",
             "PPanGGOLiN",
             "ggCaller",
             "Bifrost",
             "Cuttlefish")
)

ggplot(
  node_degree_abundance_difference,
  aes(x = degree_bin, y = frac_bin, fill = delta_pp_node)
) +
  geom_tile() +
  facet_wrap(~ Tool) +
  scale_fill_gradient2(
    name = "Δ fraction of nodes\n(draft – complete, pp)",
    low = "#2166AC",
    mid = "white",
    high = "#B2182B",
    midpoint = 0
  ) +
  scale_x_continuous(breaks = 0:10, name = "Node degree") +
  ylab("Prevalence bin") +
  labs(
    x = "Node degree",
    y = "Fraction of isolates"
  ) +
  theme_bw() +
  theme(
    axis.title.x = element_text(size = 18, face = "bold"),
    axis.text.x = element_text(size = 16),
    axis.title.y = element_text(size = 18, face = "bold"),
    axis.text.y = element_text(size = 12),
    panel.grid = element_blank(),
    strip.text = element_text(size=20,face = "bold"),
    legend.title = element_text(size = 20, face="bold",margin = margin(b = 20)), 
    legend.text = element_text(size = 14),
    legend.position = c(0.98, 0.1),
    legend.justification = c("right", "bottom")
  )+
  guides(
    fill = guide_colorbar(
      barwidth  = unit(0.8, "cm"),
      barheight = unit(4, "cm")
    )
  )


```


##3.5 Difference in node abudannce (draft vs complete) / normalize by whole graph
#Supplementary Table S7
```{r fig.width=4, fig.height=14, dpi=300}

#export a tmp table to describe the results
tmp_node_abundance_difference_per_bin<-node_degree_abundance_difference[,c(1,2,3,8,13,15)]

write.csv(tmp_node_abundance_difference_per_bin, '../results/Supplementary_table_S7.csv',  row.names = FALSE)

p_heat<-ggplot(
  node_degree_abundance_difference,
  aes(x = degree_bin, y = frac_bin, fill = delta_pp_total)
) +
  geom_tile() +
  facet_wrap(~ Tool, ncol=1) +
  scale_fill_gradient2(
    name = "Δ fraction nodes\n(draft – complete, pp)",
    low = "#2166AC",
    mid = "white",
    high = "#B2182B",
    midpoint = 0
  ) +
  scale_x_continuous(breaks = 0:10, name = "Node degree") +
  ylab("Prevalence bin") +
  labs(
    x = "Node degree",
    y = "Fraction of isolates"
  ) +
  theme_bw() +
  theme(
    axis.title.x = element_text(size = 15, face = "bold"),
    axis.text.x = element_text(size = 14),
    axis.title.y = element_text(size = 18, face = "bold"),
    axis.text.y = element_text(size = 10),
    panel.grid = element_blank(),
    strip.text = element_text(size=20,face = "bold"),
    legend.title = element_text(size = 14, face="bold",margin = margin(r = 20)), 
    legend.text = element_text(size = 14),
    legend.position = "top"
  )+
  guides(
    fill = guide_colorbar(
      barwidth  = unit(4, "cm"),
      barheight = unit(0.8, "cm")
    )
  )
p_heat
```

##3.6 Absolute Difference in node abudannce (draft vs complete) / normalize by whole graph

```{r fig.width=4, fig.height=14, dpi=300}
bin_summary <- node_degree_abundance_difference %>%
  group_by(Tool, frac_bin) %>%
  summarise(delta_pp_bin = sum(delta_pp_total), .groups = "drop")

bin_summary <- bin_summary %>%
  mutate(
    direction = case_when(
      delta_pp_bin >  0 ~ "increase",
      delta_pp_bin <  0 ~ "decrease",
      TRUE              ~ "no_change"
    )
  )

#export a tmp file for writing
#write.csv(bin_summary, '../results/TMP_bin_summary_diffrence.csv',  row.names = FALSE)


# 2. Barplot: change in fraction of nodes per prevalence bin
p_bar <- ggplot(
  bin_summary,
  aes(x = frac_bin, y = delta_pp_bin, fill = direction)
) +
  geom_col() +
  facet_wrap(~ Tool, ncol = 1) +          # match heatmap
  coord_flip() +                          # bins vertical
  # scale_fill_gradient2(
  #   name     = "Δ fraction of nodes\n(draft – complete, pp)",
  #   low      = "#2166AC",
  #   mid      = "white",
  #   high     = "#B2182B",
  #   midpoint = 0
  # ) +
  scale_fill_manual(
    values=c(
      increase="#B2182B",
      decrease="#2166AC",
      no_change="white"
    ))+
  labs(x = NULL, y = "Total Δ nodes\n (draft-complete, pp)") +
  theme_minimal() +
  theme(
    # keep strip text so it aligns with p_heat
    strip.text   = element_text(size = 20, face = "bold", colour = "white"),
    strip.background = element_rect(fill = "white", colour = NA),
    # remove y-axis label/text/ticks
    axis.title.y = element_blank(),
    axis.text.y  = element_blank(),
    axis.ticks.y = element_blank(),
    # Get the x-axys
    axis.title.x = element_text(size = 13, face = "bold"),
    axis.text.x = element_text(size = 14),
    # no legend from this panel (we'll collect from p_heat)
    legend.position = "none",
    
    #remove grid
    panel.grid.major.y = element_blank(),
    panel.grid.minor.y = element_blank(),
  )

p_bar
p_bar<-p_bar+ guides(fill = "none")
```

##3.7 Figure 4 
```{r fig.width=6, fig.height=14, dpi=300}
combined <- p_heat + p_bar +
  plot_layout(
    widths = c(6, 2),   # heatmap wider than barplot
    guides = "collect"  # share the colour legend
  ) &
  theme(
    legend.position = "top"   # legend for the whole combined figure
  )

combined
```

#4. CPU and TIme usage

##4.1. Figure 5
```{r fig.width=6,fig.height=4, dpi=300}

base <- read.csv("../../2024_01_pangenome_constructions/results/cpu_performance/Cup_time_memory.csv")


base$percentage_complete_genome <- factor(base$percentage_complete_genome, levels = c("0%", "1%", "10%", "50%", "90%", "99%", "100%"))

# only save 0%，50%，100%
base_filtered <- base %>% filter(percentage_complete_genome %in% c("0%", "50%", "100%"))

base_filtered <- base_filtered %>% filter(Tool!="pangraph")

base_long <- pivot_longer(base_filtered, cols = c(MaxRSS,Total_cpu_second), names_to = "Metric", values_to = "Value")

base_long$Tool <- recode(
  base_long$Tool,
  "panaroo"   = "Panaroo",
  "PPanGGOLiN"= "PPanGGOLiN",
  "ggcaller"  = "ggCaller",
  "bifrost"   = "Bifrost",
  "cuttlefish"= "Cuttlefish"
)

base_long_memory <- base_long %>% filter(Metric == "MaxRSS") 


ggplot(base_long_memory, aes(x = percentage_complete_genome, y = Value, colour = Tool, group = interaction(Tool, Metric))) +
  geom_line(size = 1, linetype = "dashed") +
  geom_point(size = 4) +  
  ylab('Log [Memory (Bytes)]') +
  xlab('Percentage of complete genome') +
  #facet_wrap(~ Metric, scales = "free_y") +
  scale_colour_manual(values = c(
    'PPanGGOLiN' = '#FF6600',
    'Panaroo'    = '#E00000',
    'ggCaller'   = '#D9EB00',
    'Bifrost'    = '#00BF2F',
    'Cuttlefish' = '#57A303'
  )) +
  scale_y_log10() +  
  theme_minimal() +
  theme(
    axis.line = element_line(color = "black"), 
    axis.title.x = element_text(size = 20, face="bold"),
    axis.text.x=element_text(size=18,angle=0),
    axis.title.y = element_text(size = 20, face = "bold"),
    axis.text.y=element_text(size=14,angle=0),
    legend.title= element_text(size = 18,face="bold"),
    legend.text = element_text(size = 18),
    legend.position = "none"
  ) 


base_long_cpu <- base_long %>% filter(Metric == "Total_cpu_second")

ggplot(base_long_cpu, aes(x = percentage_complete_genome, y = Value, colour = Tool, group = interaction(Tool, Metric))) +
  geom_line(size = 1, linetype = "dashed") +
  geom_point(size = 4) +  
  ylab('Log [CPU-time (s)]') +
  xlab('Percentage of complete genome') +
  #facet_wrap(~ Metric, scales = "free_y") +
  scale_colour_manual(values = c(
    'PPanGGOLiN' = '#FF6600',
    'Panaroo'    = '#E00000',
    'ggCaller'   = '#D9EB00',
    'Bifrost'    = '#00BF2F',
    'Cuttlefish' = '#57A303'
  )) +
  scale_y_log10() +  
  theme_minimal() +
  theme(
    axis.line = element_line(color = "black"), 
    axis.title.x = element_text(size = 20, face="bold"),
    axis.text.x=element_text(size=18,angle=0),
    axis.title.y = element_text(size = 20, face = "bold"),
    axis.text.y=element_text(size=14,angle=0),
    legend.title= element_text(size = 18,face="bold"),
    legend.text = element_text(size = 18),
    legend.position = "top"
  ) 


base_long_export<-base_long[,c(1,3,20,21)]
#write.csv(base_long_export, '../results/TMP_cpu_time.csv',  row.names = FALSE)

```
